% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDES_Helpers.R
\name{getModule}
\alias{getModule}
\title{Simple function to download a SpaDES module as Github repository}
\usage{
getModule(gitRepo, overwrite = FALSE, modulePath = ".")
}
\arguments{
\item{gitRepo}{A github repository, presented in the standard R way, with
\code{account/repository@branch}}

\item{overwrite}{Logical. If \code{TRUE}, then the download will delete any
existing folder with the same name as the \code{repository}
provided in \code{gitRepo}}

\item{modulePath}{A local path in which to place the full module, within
a subfolder ... i.e., the source code will be downloaded to here:
\code{file.path(modulePath, repository)}}
}
\description{
Simple function to download a SpaDES module as Github repository
}
