% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDES_Helpers.R
\name{installSpaDES}
\alias{installSpaDES}
\title{Install SpaDES packages, making sure to update.packages first}
\usage{
installSpaDES(ask = FALSE, type, libPath = .libPaths()[1])
}
\arguments{
\item{ask}{Passed to \code{update.packages}}

\item{type}{passed to both \code{update.packages} and \code{install.packages}. This
will set \code{"binary"} on windows, if not set, to get the binary packages from CRAN}

\item{libPath}{Passed to both \code{update.packages(lib.lob = libPath)} and
\code{install.packages(lib = libPath, ...)}}
}
\description{
Install SpaDES packages, making sure to update.packages first
}
