% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaDES_Helpers.R
\name{makeSureAllPackagesInstalled}
\alias{makeSureAllPackagesInstalled}
\title{Pre-test for packages in SpaDES modules}
\usage{
makeSureAllPackagesInstalled(modulePath)
}
\arguments{
\item{modulePath}{The path to modules, as per \code{SpaDES.core::setPaths}}
}
\description{
This function can be put at the start of project code. It will
only be necessary if there are multiple \code{simInit} calls (i.e., the project
doesn't use only one call). It will check all modules in \code{modulePath} for
package dependencies. It will prompt
}
\examples{
\dontrun{
source("https://raw.githubusercontent.com/PredictiveEcology/SpaDES-modules/master/R/SpaDES_Helpers.R")
out <- makeSureAllPackagesInstalled(modulePath = "modules")
}
}
